+++ tests/Support/ConfigFactory/config/events-console.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-return [
-    'e1' => [
-        ['app3', 'handler3'],
-    ],
-];